Allow accessor function to be passed in
Accessor function is option, and is used to replace the default behavior of `x, y` retrieval from entity via `entity.x` and `entity.y`.
The main driver is another project I'm working on that is storing `x` and `y` coordinates within a `position` field, so `entity.position.x` and `entity.position.y`.